requirements-dev.txt
# 开发和测试依赖

# 核心测试框架
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-html>=3.2.0
pytest-xdist>=3.3.0
pytest-timeout>=2.1.0
pytest-mock>=3.11.0
pytest-asyncio>=0.21.0
pytest-benchmark>=4.0.0

# HTTP测试工具
httpx>=0.24.0
aioresponses>=0.7.4
requests-mock>=1.11.0

# 代码质量工具
black>=23.7.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0
bandit>=1.7.5
safety>=2.3.0

# 性能测试
locust>=2.15.0
memory-profiler>=0.61.0
psutil>=5.9.0

# 数据库工具
psycopg2-binary>=2.9.7
alembic>=1.12.0
sqlalchemy-utils>=0.41.0

# 异步测试
pytest-asyncio>=0.21.0
pytest-trio>=0.8.0

# 监控和报告
pytest-sugar>=0.9.7
pytest-spec>=3.2.0
pytest-json-report>=1.5.0
coverage[toml]>=7.2.0

# 环境管理
python-dotenv>=1.0.0
tox>=4.6.0
tox-gh-actions>=3.1.0

# 文档工具
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0
mkdocs>=1.5.0

# 容器化测试
pytest-docker>=2.0.0
docker>=6.1.0

# 安全测试
semgrep>=1.35.0
OWASP ZAP API client>=1.1.0

# API文档测试
swagger-spec-validator>=3.0.0
jsonschema>=4.19.0

# 数据处理
factory-boy>=3.3.0
faker>=19.6.0

# 缓存测试
pytest-redis>=3.0.0
fakeredis>=2.18.0

# 邮件测试
pytest-mailhog>=1.0.0

# 日志测试
pytest-capturelog>=0.7.0

# 通知测试
pytest-slack>=0.3.0

# 基准测试
pytest-benchmark>=4.0.0

# 性能分析
pytest-profiling>=1.7.0
py-spy>=0.3.14

# 测试报告
allure-pytest>=2.13.0
pytest-reportlog>=0.4.0

# 并发测试
pytest-concurrent>=1.0.0

# 分布式测试
pytest-xdist>=3.3.0

# 测试数据生成
hypothesis>=6.82.0

# Mock服务
responses>=0.23.0
betamax>=0.8.1

# 测试监控
pytest-instafail>=0.5.0

# 测试插件
pytest-cookies>=1.0.0
pytest-virtualenv>=1.7.0
pytest-env>=0.8.0
pytest-doctest>=1.0.0

# 开发和调试工具
ipython>=8.14.0
ipdb>=0.13.0
pdbpp>=0.10.0

# 类型检查
types-requests>=2.31.0
types-redis>=4.6.0
types-PyYAML>=6.0.0

# 测试数据验证
cerberus>=1.3.0
voluptuous>=0.13.1

# 配置文件验证
configparser>=6.0.0
python-decouple>=3.8

# 数据库迁移测试
pytest-alembic>=0.10.0

# API契约测试
pytest-openapi>=1.0.0

# 负载测试
pytest-loadtests>=1.0.0

# 端到端测试
playwright>=1.37.0
selenium>=4.11.0

# 安全扫描
bandit[toml]>=1.7.5

# 代码复杂度分析
radon>=6.0.0
xenon>=0.9.0

# 依赖检查
pipdeptree>=2.9.0
pip-check-reqs>=2.4.0

# 测试覆盖率增强
coverage[toml]>=7.2.0
pytest-cov>=4.1.0

# 测试数据管理
pytest-fixtures>=0.1.0
test-data-loader>=1.0.0

# 并发测试工具
pytest-parallel>=0.1.0
pytest-trio>=0.8.0

# 监控测试
prometheus-client>=0.17.0
pytest-prometheus>=1.0.0

# 性能监控
pytest-perf>=0.3.0

# 日志测试增强
pytest-capturelog>=0.7.0
structlog>=23.1.0

# API测试增强
pytest-httpx>=0.27.0

# 异步测试增强
pytest-trio>=0.8.0
pytest-aioresponses>=0.7.0

# 数据库测试增强
pytest-sqlalchemy>=0.2.0

# 配置测试
pytest-config>=0.1.0
